Fixed the wrong address being assigned to suit_arr
At lines 103, 108 and 119 I was assigning the address of the pointer itself, not the address to which it points (the PlayingCard struct). You bumbling dummy.

I wonder if those breaks are necessarry for thing to work fine now, I might see about that later.

The function now works as intended for the first case. Onto the others.