Broke and fixed the whole project
In the hope of finding out why the unit test project works fine on Debug and has a segmentation fault on Build & Run, I discovered AddressSanitizer.

Naively, I just added whatever compilator option I saw and thought this would be a panacea to my problems. Little did I know this made things even worse as the project no longer compiled. I was flooded with missing references to asan library.

I'm adding this to the notes.c as a caution for me in the future.