Various fixes & debug prints
Added a handful of debugging prints.
Fixed turns_left not being reset back to player_count - 1 when beginning a new betting round.
Fixed players not being marked as AI.
Fixed the gameOver not being negated in the while loop.
That unit-test was equaling the actual value to the wrong expected value.