Fixed dealer button being assigned to folded players
gamestate.passDealerButton did not care about player's folded property. In order to access it now, signature of that function was changed.

As of right now, playerio gives blind status higher priority than being a dealer when it comes to printing the status of the game. Nothing needed to be changed there.